backend/requirements.txt
fastapi[standard]
sqlmodel
psycopg2-binary
passlib[bcrypt]
pyjwt
python-dotenv
pydantic-settings
argon2_cffi
pydantic[email]

fastapi-sqlalchemy
pydantic
alembic
uvicorn

pytest
pytest-asyncio
httpx
pytest-mock